Move modals to modals folder
